moved link, image, video and image preview outside flat list
