+++ src/components/postComments/container/postComments.tsx
@@ -5,11 +5,14 @@ import React, {
   useState,
   useMemo,
   useEffect,
+  Fragment,
 } from 'react';
-import { ActivityIndicator, Platform, RefreshControl, Text } from 'react-native';
+import { ActivityIndicator, Alert, Modal, PermissionsAndroid, Platform, RefreshControl, Text } from 'react-native';
 import { useIntl } from 'react-intl';
 import { useNavigation } from '@react-navigation/native';
 import { FlatList } from 'react-native-gesture-handler';
+import ActionsSheet from 'react-native-actions-sheet';
+import ImageViewer from 'react-native-image-zoom-viewer';
 
 // Components
 import { postBodySummary } from '@ecency/render-helper';
@@ -27,6 +30,10 @@ import { CacheStatus } from '../../../redux/reducers/cacheReducer';
 import { CommentsSection } from '../children/commentsSection';
 import styles from '../children/postComments.styles';
 import { PostTypes } from '../../../constants/postTypes';
+import CameraRoll from '@react-native-community/cameraroll';
+import RNFetchBlob from 'rn-fetch-blob';
+import { OptionsModal } from '../../atoms';
+import VideoPlayer from '../../videoPlayer/videoPlayerView';
 
 
 const PostComments = forwardRef(
@@ -57,11 +64,21 @@ const PostComments = forwardRef(
 
     const writeCommentRef = useRef(null);
     const commentsListRef = useRef<FlatList | null>(null);
+    const actionImage = useRef(null);
+    const actionLink = useRef(null);
+    const youtubePlayerRef = useRef(null);
 
+    const [postImages, setPostImages] = useState<string[]>([]);
+    const [isImageModalOpen, setIsImageModalOpen] = useState(false);
     const [selectedFilter, setSelectedFilter] = useState('trending');
     const [selectedOptionIndex, setSelectedOptionIndex] = useState(0);
     const [shouldRenderComments, setShouldRenderComments] = useState(false);
     const [headerHeight, setHeaderHeight] = useState(0);
+    const [videoUrl, setVideoUrl] = useState(null);
+    const [youtubeVideoId, setYoutubeVideoId] = useState(null);
+    const [videoStartTime, setVideoStartTime] = useState(0);
+    const [selectedImage, setSelectedImage] = useState(null);
+    const [selectedLink, setSelectedLink] = useState(null);
 
     const sortedSections = useMemo(
       () => _sortComments(selectedFilter, discussionQuery.sectionedData),
@@ -95,6 +112,88 @@ const PostComments = forwardRef(
       onRefresh();
     };
 
+    const checkAndroidPermission = async () => {
+      try {
+        const permission = PermissionsAndroid.PERMISSIONS.WRITE_EXTERNAL_STORAGE;
+        await PermissionsAndroid.request(permission);
+        Promise.resolve();
+      } catch (error) {
+        Promise.reject(error);
+      }
+    };
+
+    const _downloadImage = async (uri) => {
+      return RNFetchBlob.config({
+        fileCache: true,
+        appendExt: 'jpg',
+      })
+        .fetch('GET', uri)
+        .then((res) => {
+          const { status } = res.info();
+
+          if (status == 200) {
+            return res.path();
+          } else {
+            Promise.reject();
+          }
+        })
+        .catch((errorMessage) => {
+          Promise.reject(errorMessage);
+        });
+    };
+
+    const _saveImage = async (uri) => {
+      try {
+        if (Platform.OS === 'android') {
+          await checkAndroidPermission();
+          uri = `file://${await _downloadImage(uri)}`;
+        }
+        CameraRoll.saveToCameraRoll(uri)
+          .then(() => {
+            dispatch(
+              toastNotification(
+                intl.formatMessage({
+                  id: 'post.image_saved',
+                }),
+              ),
+            );
+          })
+          .catch(() => {
+            dispatch(
+              toastNotification(
+                intl.formatMessage({
+                  id: 'post.image_saved_error',
+                }),
+              ),
+            );
+          });
+      } catch (error) {
+        dispatch(
+          toastNotification(
+            intl.formatMessage({
+              id: 'post.image_saved_error',
+            }),
+          ),
+        );
+      }
+    };
+
+    const _handleYoutubePress = (videoId, startTime) => {
+      if (videoId && youtubePlayerRef.current) {
+        setYoutubeVideoId(videoId);
+        setVideoStartTime(startTime);
+        youtubePlayerRef.current.setModalVisible(true);
+      }
+    };
+
+    const _handleVideoPress = (embedUrl) => {
+      if (embedUrl && youtubePlayerRef.current) {
+        setVideoUrl(embedUrl);
+        setVideoStartTime(0);
+        youtubePlayerRef.current.setModalVisible(true);
+      }
+    };
+
 
     const _handleOnDropdownSelect = (option, index) => {
       setSelectedFilter(option);
@@ -109,7 +208,7 @@ const PostComments = forwardRef(
           content,
         },
         key: content.permlink,
-      });
+      } as never);
     };
 
     const _handleOnEditPress = (item) => {
@@ -121,7 +220,7 @@ const PostComments = forwardRef(
           isReply: true,
           post: item,
         },
-      });
+      } as never);
     };
 
     const _handleDeleteComment = (_permlink) => {
@@ -197,6 +296,77 @@ const PostComments = forwardRef(
       );
     };
 
+
+
+    const _handleImageOptionPress = (ind) => {
+      if (ind === 1) {
+        // open gallery mode
+        setIsImageModalOpen(true);
+      }
+      if (ind === 0) {
+        // copy to clipboard
+        writeToClipboard(selectedImage).then(() => {
+          dispatch(
+            toastNotification(
+              intl.formatMessage({
+                id: 'alert.copied',
+              }),
+            ),
+          );
+        });
+      }
+      if (ind === 2) {
+        // save to local
+        _saveImage(selectedImage);
+      }
+
+      setSelectedImage(null);
+    };
+
+    const _handleLinkOptionPress = (ind) => {
+      if (ind === 1) {
+        // open link
+        if (selectedLink) {
+          navigation.navigate({
+            name: ROUTES.SCREENS.WEB_BROWSER,
+            params: {
+              url: selectedLink,
+            },
+            key: selectedLink,
+          } as never);
+        }
+      }
+      if (ind === 0) {
+        // copy to clipboard
+        writeToClipboard(selectedLink).then(() => {
+          dispatch(
+            toastNotification(
+              intl.formatMessage({
+                id: 'alert.copied',
+              }),
+            ),
+          );
+        });
+      }
+
+      setSelectedLink(null);
+    };
+
+
+
+    const _handleLinkPress = (url:string) => {
+     
+      setSelectedLink(url);
+      actionLink.current?.show();
+    }
+
+    const _handleImagePress = (url: string, postImgUrls: string[]) => {
+      setPostImages(postImgUrls);
+      setSelectedImage(url);
+      actionImage.current?.show();
+    }
+
+
     const _onContentSizeChange = (x: number, y: number) => {
       // lazy render comments after post is rendered;
       if (!shouldRenderComments) {
@@ -243,44 +413,114 @@ const PostComments = forwardRef(
     };
 
     const _renderItem = ({ item, index }) => {
-      return <CommentsSection
-        item={item}
-        index={index}
-        mainAuthor={mainAuthor}
-        handleDeleteComment={_handleDeleteComment}
-        handleOnEditPress={_handleOnEditPress}
-        handleOnVotersPress={_handleOnVotersPress}
-        handleOnLongPress={_handleShowOptionsMenu}
-        handleOnUserPress={_handleOnUserPress}
-        openReplyThread={_openReplyThread}
-        onUpvotePress={(anchorRect, showPayoutDetails)=>onUpvotePress({anchorRect, showPayoutDetails, content:item, postType:PostTypes.COMMENT})}
-      />
+      return (
+        <CommentsSection
+          item={item}
+          index={index}
+          mainAuthor={mainAuthor}
+          handleDeleteComment={_handleDeleteComment}
+          handleOnEditPress={_handleOnEditPress}
+          handleOnVotersPress={_handleOnVotersPress}
+          handleOnLongPress={_handleShowOptionsMenu}
+          handleOnUserPress={_handleOnUserPress}
+          handleImagePress={_handleImagePress}
+          handleLinkPress={_handleLinkPress}
+          handleVideoPress={_handleVideoPress}
+          handleYoutubePress={_handleYoutubePress}
+          openReplyThread={_openReplyThread}
+          onUpvotePress={(anchorRect, showPayoutDetails) => onUpvotePress({ anchorRect, showPayoutDetails, content: item, postType: PostTypes.COMMENT })}
+        />
+      )
     }
 
 
 
     return (
-      <FlatList
-        ref={commentsListRef}
-        style={styles.list}
-        contentContainerStyle={styles.listContent}
-        ListHeaderComponent={_postContentView}
-        ListEmptyComponent={_renderEmptyContent}
-        data={shouldRenderComments ? sortedSections : []}
-        onContentSizeChange={_onContentSizeChange}
-        renderItem={_renderItem}
-        keyExtractor={(item) => `${item.author}/${item.permlink}`}
-        refreshControl={
-          <RefreshControl
-            refreshing={discussionQuery.isFetching}
-            onRefresh={_onRefresh}
-            progressBackgroundColor="#357CE6"
-            tintColor={!isDarkTheme ? '#357ce6' : '#96c0ff'}
-            titleColor="#fff"
-            colors={['#fff']}
+      <Fragment>
+        <FlatList
+          ref={commentsListRef}
+          style={styles.list}
+          contentContainerStyle={styles.listContent}
+          ListHeaderComponent={_postContentView}
+          ListEmptyComponent={_renderEmptyContent}
+          data={shouldRenderComments ? sortedSections : []}
+          onContentSizeChange={_onContentSizeChange}
+          renderItem={_renderItem}
+          keyExtractor={(item) => `${item.author}/${item.permlink}`}
+          refreshControl={
+            <RefreshControl
+              refreshing={discussionQuery.isFetching}
+              onRefresh={_onRefresh}
+              progressBackgroundColor="#357CE6"
+              tintColor={!isDarkTheme ? '#357ce6' : '#96c0ff'}
+              titleColor="#fff"
+              colors={['#fff']}
+            />
+          }
+        />
+
+
+        <Modal visible={isImageModalOpen} transparent={true}>
+          <ImageViewer
+            imageUrls={postImages.map((url) => ({ url }))}
+            enableSwipeDown
+            onCancel={() => setIsImageModalOpen(false)}
+            onClick={() => setIsImageModalOpen(false)}
           />
-        }
-      />
+        </Modal>
+
+
+        <OptionsModal
+          ref={actionImage}
+          options={[
+            intl.formatMessage({ id: 'post.copy_link' }),
+            intl.formatMessage({ id: 'post.gallery_mode' }),
+            intl.formatMessage({ id: 'post.save_to_local' }),
+            intl.formatMessage({ id: 'alert.cancel' }),
+          ]}
+          title={intl.formatMessage({ id: 'post.image' })}
+          cancelButtonIndex={3}
+          onPress={(index) => {
+            _handleImageOptionPress(index);
+          }}
+        />
+
+
+        <OptionsModal
+          ref={actionLink}
+          options={[
+            intl.formatMessage({ id: 'post.copy_link' }),
+            intl.formatMessage({ id: 'alert.external_link' }),
+            intl.formatMessage({ id: 'alert.cancel' }),
+          ]}
+          title={intl.formatMessage({ id: 'post.link' })}
+          cancelButtonIndex={2}
+          onPress={(index) => {
+            _handleLinkOptionPress(index);
+          }}
+        />
+
+        <ActionsSheet
+          ref={youtubePlayerRef}
+          gestureEnabled={true}
+          hideUnderlay={true}
+          containerStyle={{ backgroundColor: 'black' }}
+          indicatorColor={EStyleSheet.value('$primaryWhiteLightBackground')}
+          onClose={() => {
+            setYoutubeVideoId(null);
+            setVideoUrl(null);
+          }}
+        >
+          <VideoPlayer
+            mode={youtubeVideoId ? 'youtube' : 'uri'}
+            youtubeVideoId={youtubeVideoId}
+            uri={videoUrl}
+            startTime={videoStartTime}
+          />
+        </ActionsSheet>
+
+      </Fragment>
+
     );
   },
 );
@@ -333,7 +573,7 @@ const _sortComments = (sortOrder = 'trending', _comments) => {
       return 0;
     },
     votes: (a, b) => {
-      if (a.renderOnTop){
+      if (a.renderOnTop) {
         return -1;
       }
 
@@ -350,7 +590,7 @@ const _sortComments = (sortOrder = 'trending', _comments) => {
       return 0;
     },
     age: (a, b) => {
-      if (a.renderOnTop){
+      if (a.renderOnTop) {
         return -1;
       }
